remove brush and replace with time button chart
