+++ game/js/src/components/charts/components/timebuttons/style.js
@@ -1,0 +1,12 @@
+import styled from 'styled-components';
+
+import { Button } from '@Style'
+
+export const TimeButtonWrapper = styled.div`
+  display: flex;
+  margin: 1rem 0;
+`;
+
+export const TimeButton = styled(Button)`
+  flex-grow: 1;
+`
