+++ game/js/src/components/charts/index.js
@@ -5,6 +5,7 @@ import { select } from 'd3-selection';
 
 import { writeGraph, reducer, init } from './components/reducers/line'
 import { writeGraph as writeBrushGraph, reducer as brushReducer, init as brushInit } from './components/reducers/brush'
+import { writeGraph as writeTimeGraph, reducer as timeReducer, init as timeInit } from './components/reducers/time';
 import { TimeButtons } from './components/timebuttons';
 
 export const LineChart = ({
@@ -73,7 +74,7 @@ LineChart.propTypes = {
   rootId: PropTypes.string.isRequired,
 };
 
-export const LineChartWithBrush = ({
+export const LineChartWithTimeButtons = ({
   xValues,
   yValues,
   labels,
@@ -95,7 +96,7 @@ export const LineChartWithBrush = ({
   const [
     state, dispatch,
   ] = useReducer(
-      brushReducer, initState, brushInit,
+      timeReducer, initState, timeInit,
   );
 
   const {
@@ -123,7 +124,7 @@ export const LineChartWithBrush = ({
         .attr('id', `${rootWrapper}`)
         .attr('transform', `translate(${margin.left}, ${margin.top})`);
 
-    writeBrushGraph(state, dispatch);
+    writeTimeGraph(state, dispatch);
   }, [
   ]);
 
@@ -136,7 +137,7 @@ export const LineChartWithBrush = ({
   );
 };
 
-LineChartWithBrush.propTypes = {
+LineChartWithTimeButtons.propTypes = {
   xValues: PropTypes.array.isRequired,
   yValues: PropTypes.array.isRequired,
   labels: PropTypes.arrayOf(PropTypes.string),
